Able to add new product to database and display users from admin panel

+++ WebProjElective/Models/Product.cs
@@ -6,7 +6,7 @@
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
         public byte[] ProductImage { get; set; }
-        public int ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public int ProductAvailableItems { get; set; }
         public DateTime ProductDateUpload { get; set; }
